Break out Progress into components per section
